closes #330 - refactor code, branched AdminLessons component out to more files
